feat(#138): move common classes to upper level
