+++ src/main/java/org/eolang/jeo/representation/DefaultVersion.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.jeo.representation.asm;
+package org.eolang.jeo.representation;
 
 import org.objectweb.asm.Opcodes;
