+++ src/main/java/org/eolang/jeo/representation/directives/DirectivesClass.java
@@ -27,7 +27,7 @@ import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Iterator;
-import org.eolang.jeo.representation.asm.DefaultVersion;
+import org.eolang.jeo.representation.DefaultVersion;
 import org.objectweb.asm.ClassVisitor;
 import org.objectweb.asm.MethodVisitor;
 import org.xembly.Directive;
