+++ src/test/java/org/eolang/jeo/representation/EoRepresentationTest.java
@@ -24,7 +24,7 @@
 package org.eolang.jeo.representation;
 
 import com.jcabi.matchers.XhtmlMatchers;
-import org.eolang.jeo.representation.asm.Bytecode;
+import org.eolang.jeo.representation.generation.Bytecode;
 import org.eolang.jeo.representation.generation.BytecodeClass;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
